Materials for #66, #67, and another issue
See info.txt in alsoneeded directory for more information on this.

These files need to replace the standard versions of them in order to
fix #66, #67, and an issue with eeprom_is_ready() in eeprom.h that
manifests on the newer parts. Unfortunately I'm not good enough at shell
scripting to make the build scripts slip these fixes in